Created useColorGeneration hook.
