+++ src/components/CardGrid.tsx
@@ -1,8 +1,9 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import ColorSelect from './ColorSelect';
 import ContrastSelect from './ContrastSelect';
 import ColorPickerList from './ColorPickerList';
 import { generateAdditionalColors } from '../utils';
+import useColorGeneration from '../hooks/useColorGeneration';
 import RadioButtonList from './RadioButtonList';
 import SectionTitle from './SectionTitle';
 import PaletteHandler from './PaletteHandler';
@@ -11,20 +12,12 @@ const CardGrid: React.FC = () => {
     // You need to have two colors to check contrast
     const [selectedOption, setSelectedOption] = useState(2);
     const [selectedContrast, setSelectedContrast] = useState(4.5);
-    const [colors, setColors] = useState<string[]>([]);
-    const [visibleColors, setVisibleColors] = useState<number>(0);
     const [paletteType, setPaletteType] = useState('Adjacent');
     const [checkboxChecked, setCheckboxChecked] = useState(false);
 
-    // TODO: Check if this can be done in state management
-    const updateVisibleColors = (newVisibleColors: number): void => {
-        if (newVisibleColors < 2) {
-            // eslint-disable-next-line no-console
-            console.error('visibleColors cannot be less than 2 ');
-            return;
-        }
-        setVisibleColors(newVisibleColors);
-    };
+    const { colors, visibleColors, setColors } =
+        useColorGeneration(selectedOption);
+
     const handleColorChange = (newColors: string[]): void => {
         setColors(newColors);
     };
@@ -33,17 +26,6 @@ const CardGrid: React.FC = () => {
         const newColors = generateAdditionalColors(visibleColors);
         setColors(newColors);
     };
-
-    useEffect(() => {
-        if (selectedOption > colors.length) {
-            // If selectedOption increases, generate new colors for the additional color pickers
-            const additionalColors = generateAdditionalColors(
-                selectedOption - colors.length,
-            );
-            setColors((prevColors) => [...prevColors, ...additionalColors]);
-        }
-        updateVisibleColors(selectedOption);
-    }, [colors.length, selectedOption]);
 
     return (
         <div className="max-w-full mx-auto px-4 sm:px-6 lg:px-8">
